refactor(react): card 2.0
