+++ packages/react/src/components/card/InteractiveCard/InteractiveCard.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import styled from 'styled-components';
-import { Card, CardProps, Box } from '../../../../layout';
+import { Card, CardProps } from '../Card';
+import { Box } from '../../../layout';
 
 export interface InteractiveCardProps extends Omit<CardProps, 'elevation' | 'position' | 'border' | 'borderColor'> {
   /** Additional action buttons that appear on hover. */
