+++ packages/react/src/components/card/README.md
@@ -4,6 +4,8 @@ Renders a card based on the elevation level.
 
 ## Usage
 
+### `<Card />`
+
 To use this component in your app, import as follows:
 
 ```jsx
@@ -47,3 +49,34 @@ export default function Component({ children }) {
   );
 }
 ```
+
+### `<InteractiveCard />`
+
+To use this component in your app, import as follows:
+
+```tsx
+import { InteractiveCard, Box, Text, Heading } from '@aksara-ui/react';
+
+export default function Component() {
+  return (
+    <InteractiveCard>
+      <Box p="lg">
+        <Heading scale={400}>Deno</Heading>
+      </Box>
+      <Box p="lg" pt={0}>
+        <Text scale={200} color="grey07">
+          A secure runtime for JavaScript and TypeScript
+        </Text>
+      </Box>
+    </InteractiveCard>
+  );
+}
+```
+
+#### Props
+
+| Property | Type            | Required | Default   | Description                                     |
+| -------- | --------------- | -------- | --------- | ----------------------------------------------- |
+| actions  | React.ReactNode | -        | undefined | Additional action buttons that appear on hover. |
+| image    | string          | -        | undefined | Any image to add to the card.                   |
+| imageAlt | string          | -        | undefined | Alt text to the card image.                     |
