+++ packages/react/src/components/card/index.ts
@@ -1,1 +1,4 @@
+export * from './Card';
+export * from './CardFooter';
 export * from './CardHeader';
+export * from './InteractiveCard';
