+++ packages/react/src/components/dropdown/components/DropdownMenu/DropdownMenu.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
-import { Box, BoxProps, Card } from '../../../../layout';
-import { Arrow } from './styles';
+import { Box, BoxProps } from '../../../../layout';
 
 export interface DropdownMenuProps extends BoxProps {
   className?: string;
@@ -11,11 +10,21 @@ export interface DropdownMenuProps extends BoxProps {
 
 const MenuArrow: React.FC<Pick<DropdownMenuProps, 'tipOffset'>> = ({ tipOffset }) => {
   return (
-    <Arrow style={{ left: tipOffset, textAlign: 'left' }}>
+    <Box
+      sx={{
+        display: 'block',
+        lineHeight: '11px',
+        zIndex: 1,
+        textAlign: 'center',
+        position: 'absolute',
+        top: '-1px',
+      }}
+      style={{ left: tipOffset, textAlign: 'left' }}
+    >
       <svg width="24" height="12" viewBox="0 0 24 12">
         <path fill="#fff" fillRule="evenodd" d="M20 12l-8-8-12 12" />
       </svg>
-    </Arrow>
+    </Box>
   );
 };
 
@@ -26,18 +35,19 @@ const DropdownMenu: React.ForwardRefRenderFunction<HTMLDivElement, DropdownMenuP
   return (
     <Box ref={ref} display="inline-block" position="relative" {...rest}>
       {tipOffset && <MenuArrow tipOffset={tipOffset} />}
-      <Card
-        elevation={3}
+      <Box
         display="inline-block"
         textAlign="left"
         width={width}
         maxWidth="100vw"
         marginTop={11}
         borderRadius="xs"
+        overflow="hidden"
+        boxShadow={3}
         py="xs"
       >
         {children}
-      </Card>
+      </Box>
     </Box>
   );
 };
