--- packages/react/src/components/interactive-card/components/index.ts
@@ -1,1 +1,0 @@
-export * from './InteractiveCard';
