--- packages/react/src/components/interactive-card/index.ts
@@ -1,1 +1,0 @@
-export * from './components';
