+++ packages/react/src/components/notification/components/Toaster.tsx
@@ -4,7 +4,8 @@ import { variant } from 'styled-system';
 import { Transition } from 'react-transition-group';
 import { IconTickSingle, IconWarningTriangle, IconCloseRounded } from '@aksara-ui/icons';
 
-import { Text, Card, Heading, Box } from '../../../layout';
+import { Card } from '../../card';
+import { Text, Heading, Box } from '../../../layout';
 import { ToasterSettings, NotificationStatus } from '../utils/types';
 import { ANIMATION_DURATION, TOASTER_WIDTH } from '../utils/constants';
 import { toasterVariants, toasterTextVariants } from './variants';
