+++ packages/react/src/components/notification/index.ts
@@ -1,5 +1,8 @@
 import Notification from './Notification';
 
+/**
+ * @deprecated Please use `useToast()` instead.
+ */
 const notification = new Notification();
 
 export { notification };
