--- packages/react/src/layout/card/index.ts
@@ -1,1 +1,0 @@
-export * from './components';
