+++ packages/react/src/layout/index.ts
@@ -1,5 +1,4 @@
 export * from './box';
-export * from './card';
 export * from './inline';
 export * from './stack';
 export * from './typography';
